// src/app/models/member-payload.ts
export interface MemberPayload {
  perNr: number;
  nume: string;
  dataAngajarii: string;
  email: string;
  dataNasterii: string;
  gen: string;
  oras: string;
  departament: string;
  businessUnit: string;
  norma: number;
  fte: number;
  formaColaborare: string;
  tipContract: string;
  functie: string;
  dreptConcediu: number;
  hrManager: string;
  project: string;
  projectStartDate: string;
  projectEndDate: string;
  client: string;
  projectManager: string;
  german: string;
  english: string;
  gLevel: string;
  skills: string;
  photoUrl: string;
}